chore(actions): manually trigger release after api update [skip ci]
Note: release check is triggered on push, but pushing from within
another github action won't trigger a push (when using the existing
token).  Alternatively we could create a personal access token
(PAT) and push with that.  However, those expire and are a pain to
manage.  So let's rather just manually trigger the release after
push in this case.